requirements.txt
keras==2.6.0
matplotlib==3.5.0
numpy==1.19.5
opencv_python==4.5.3.56
Pillow==9.1.0
seaborn==0.11.2
streamlit==1.8.1
tensorflow==2.8.0
tensorflow_gpu==2.6.0
